prepare components, NOT entegrated
