+++ src/components/NewsList.js
@@ -1,0 +1,21 @@
+/** @jsx createElement */
+/** @jsxFrag createFragment */
+import { createElement, createFragment } from '../framework/element';
+import { normalizeNews } from '../data/dataHandlers';
+import { NewsItem } from './NewsItem';
+
+export function NewsList(news) {
+  return `
+  ${
+    news === undefined
+      ? `<h3>No matches for your search</h3>`
+      : normalizeNews(news).length > 0
+      ? normalizeNews(news)
+          .map(item => {
+            return NewsItem(item);
+          })
+          .join('')
+      : `<h3>No matches for your search</h3>`
+  }
+  `;
+}
